CPG-39 CPG-34 creating schema and Footer Component structure
