+++ src/app/globals.css
@@ -1,20 +1,3 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-html {
-    scroll-behavior: smooth;
-}
-
-ul {
-    list-style: none;
-}
-li * {
-    color: inherit;
-    text-decoration: none;
-}
-
 :root {
     --header-height:  80px;
     --header-height-scrolled: 70px;
@@ -45,7 +28,19 @@ li * {
     --font-size-2xl: 1.5rem;
 }
 
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}    
+html {
+    scroll-behavior: smooth;
+}    
 
-
-.btn-primary {}
-.btn-secondary {}
+ul {
+    list-style: none;
+}    
+li * {
+    color: inherit;
+    text-decoration: none;
+}
